requirements.txt
argon2-cffi==19.1.0
asgiref==3.2.10
astroid==2.4.2
backports.csv==1.0.7
behave==1.2.6
behave-django==1.1.0
cffi==1.14.0
Click==7.0
colorama==0.4.3
csv-2-mongo==0.0.2
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.2
django-adaptors==0.2.5
django-bootstrap-modal-forms==1.3.1
django-cors-headers==3.1.0
django-cors-middleware==1.4.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.1.0
django-heroku==0.3.1
django-import-export==1.2.0
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.3.1
django-settings-json==1.0.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djongo==1.2.33
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==4.1.1
Flask==1.1.1
Flask-Login==0.4.1
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
numpy==1.16.4
odfpy==1.4.0
opencv-contrib-python==4.1.2.30
openpyxl==2.5.12
packaging==19.0
pandas==0.25.0
parse==1.15.0
parse-type==0.5.2
Pillow==6.1.0
pip-review==1.0
ply==3.11
psycopg2==2.8.5
pycparser==2.20
pygame==1.9.6
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
reportlab==3.5.12
six==1.15.0
sqlparse==0.2.4
tablib==0.13.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
virtualenv==16.6.2
virtualenvwrapper-win==1.2.5
Werkzeug==0.15.5
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
